+++ CHANGELOG.md
@@ -10,7 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 - Clicking close button on tab stack now destroys it.
 - Fixed context menu actions for tab stacks not triggering.
-
+- Fixed layout manager not cleaning up empty stacks when closing all tabs.
 ### Added
 - New "90's Workstation" (IRIX/CDE style) theme.